requirements.txt
arrow==1.2.2
asgiref==3.5.2
autopep8==1.6.0
binaryornot==0.4.4
campay==1.0.3
certifi==2022.5.18.1
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.5
django-autoslug==1.9.8
django-browser-reload==1.5.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-tailwind==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.1.1
poyo==0.5.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==4.2.0
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
